[generator][search] Move postcode_points section build to separate file.
