+++ generator/search_index_builder.hpp
@@ -2,11 +2,6 @@
 
 #include <string>
 
-namespace storage
-{
-class CountryInfoGetter;
-}
-
 namespace indexer
 {
 // Builds the latest version of the search index section and writes it to the mwm file.
@@ -15,12 +10,4 @@ namespace indexer
 // in version mismatch when trying to read the index.
 bool BuildSearchIndexFromDataFile(std::string const & path, std::string const & country,
                                   bool forceRebuild, uint32_t threadsCount);
-
-// Builds postcodes section with external postcodes data and writes it to the mwm file.
-bool BuildPostcodes(std::string const & path, std::string const & country,
-                    std::string const & datasetPath, bool forceRebuild);
-// Exposed for testing.
-bool BuildPostcodesWithInfoGetter(std::string const & path, std::string const & country,
-                                  std::string const & datasetPath, bool forceRebuild,
-                                  storage::CountryInfoGetter & infoGetter);
 }  // namespace indexer
